Fix CDROM 64-bit crash (Alexander van Heukelum, SF#3228939).
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@2016 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/dosext/mfs/mscdex.c
@@ -75,7 +75,7 @@ static int ReadSectors(int drive, Bit32u sector, Bit16u num, unsigned char *buf)
 	Bit8u driver = GetDriver(drive);
 	if (driver >= 4)
 		return MSCDEX_ERROR_UNKNOWN_DRIVE;
-	*CALC_PTR(req_buf, MSCD_READ_STARTSECTOR, u_long) = sector;
+	*CALC_PTR(req_buf, MSCD_READ_STARTSECTOR, u_int) = sector;
 	*CALC_PTR(req_buf, MSCD_READ_ADRESSING, u_char) = 0;
 	*CALC_PTR(req_buf, MSCD_READ_NUMSECTORS, u_short) = num;
 	_AH = 2 | (driver << 6);
